Moved messages to dedicated assembly
