--- src/Features/MetadataProcessor/Events/ProcessMetadataRequest.cs
@@ -1,3 +1,0 @@
-namespace Kurmann.Videoschnitt.Features.MetadataProcessor.Events;
-
-public record ProcessMetadataRequest {}
