+++ src/Messages/Metadata/ProcessMetadataRequest.cs
@@ -1,0 +1,3 @@
+namespace Kurmann.Videoschnitt.Messages.Metadata;
+
+public record ProcessMetadataRequest {}
